feat: refactor transaction handling and formatting for improved clarity and consistency
